tests/pyproject.toml
[project]
name = "tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.3.5",
    "pytest-html>=4.1.1",
    "pytest-order>=1.3.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
